+++ ComputerShop/AdminWindow/Add/AddUserPage.xaml
@@ -1,0 +1,125 @@
+﻿<Window x:Class="Dashboard.AdminWindow.Add.AddUserPage"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:Dashboard.AdminWindow.Add"
+        xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
+        mc:Ignorable="d"
+        Title="Thêm Nhân Viên"
+        Height="auto"
+        Width="auto"
+        SizeToContent="WidthAndHeight"
+        WindowStartupLocation="CenterScreen">
+    <Grid>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition></ColumnDefinition>
+            <ColumnDefinition></ColumnDefinition>
+        </Grid.ColumnDefinitions>
+        <GroupBox x:Name="gb_image"
+                  Header="Card Group Box"
+                  Style="{DynamicResource MaterialDesignCardGroupBox}"
+                  Margin="16"
+                  MouseDown="GroupBox_MouseDown"
+                  Drop="GroupBox_Drop"
+                  MaxHeight="300"
+                  MaxWidth="300"
+                  MinHeight="180"
+                  AllowDrop="True"
+                  VerticalAlignment="Top">
+            <GroupBox.HeaderTemplate>
+                <DataTemplate>
+                    <StackPanel Orientation="Horizontal">
+                        <materialDesign:PackIcon Kind="ImageArea"
+                                                 Height="32"
+                                                 Width="32"
+                                                 VerticalAlignment="Center" />
+                        <TextBlock Margin="8,0,0,0"
+                                   VerticalAlignment="Center"
+                                   Style="{StaticResource MaterialDesignSubtitle1TextBlock}"
+                                   Text="{Binding}" />
+                    </StackPanel>
+                </DataTemplate>
+            </GroupBox.HeaderTemplate>
+            <Image x:Name="img_Product"
+                   VerticalAlignment="Center"
+                   HorizontalAlignment="Center" />
+        </GroupBox>
+
+        <StackPanel x:Name="stackPanel"
+                    Grid.Column="1"
+                    HorizontalAlignment="Left"
+                    VerticalAlignment="Top"
+                    Width="700"
+                    Margin="10"
+                    Background="White">
+            <TextBox x:Name="tbx_ID"
+                     materialDesign:HintAssist.Hint="ID"
+                     Style="{StaticResource MaterialDesignFloatingHintTextBox}"
+                     VerticalAlignment="Center" />
+            <TextBox x:Name="tbx_FirstName"
+                     materialDesign:HintAssist.Hint="Tên"
+                     Style="{StaticResource MaterialDesignFloatingHintTextBox}"
+                     VerticalAlignment="Center" />
+            <TextBox x:Name="tbx_LastName"
+                     materialDesign:HintAssist.Hint="Last Name"
+                     Style="{StaticResource MaterialDesignFloatingHintTextBox}"
+                     VerticalAlignment="Center" />
+            <Grid>
+
+                <DatePicker x:Name="dp_DoB"
+                            materialDesign:HintAssist.Hint="Pick Date"
+                            Style="{StaticResource MaterialDesignFloatingHintDatePicker}" />
+                <Button x:Name="btn_OpenCalendar"
+                        Height="40"
+                        Click="btn_OpenCalendar_Click"
+                        Background="#00673AB7"
+                        BorderBrush="#00673AB7" />
+            </Grid>
+
+            <TextBox x:Name="tbx_Email"
+                     materialDesign:HintAssist.Hint="Email"
+                     Style="{StaticResource MaterialDesignFloatingHintTextBox}"
+                     VerticalAlignment="Center" />
+            <TextBox x:Name="tbx_PhoneNumber"
+                     materialDesign:HintAssist.Hint="PhoneNumber"
+                     Style="{StaticResource MaterialDesignFloatingHintTextBox}"
+                     VerticalAlignment="Center"
+                     PreviewTextInput="NumberValidationTextBox" />
+            <TextBox x:Name="tbx_Username"
+                     materialDesign:HintAssist.Hint="User Name"
+                     Style="{StaticResource MaterialDesignFloatingHintTextBox}"
+                     VerticalAlignment="Center" />
+            <TextBox x:Name="tbx_Password"
+                     materialDesign:HintAssist.Hint="Password"
+                     Style="{StaticResource MaterialDesignFloatingHintTextBox}"
+                     VerticalAlignment="Center" />
+            <TextBox x:Name="tbx_ConfirmPassword"
+                     materialDesign:HintAssist.Hint="Confirm Password"
+                     Style="{StaticResource MaterialDesignFloatingHintTextBox}"
+                     VerticalAlignment="Center" />
+            <StackPanel Orientation="Horizontal"
+                        Height="40">
+                <ToggleButton x:Name="tbtn_IsAdmin"
+                              Style="{StaticResource MaterialDesignSwitchToggleButton}"
+                              ToolTip="Default ToggleButton Style"
+                              Width="50"
+                              Height="40" />
+                <TextBlock VerticalAlignment="Center">Admin</TextBlock>
+            </StackPanel>
+            <Grid Height="40">
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition></ColumnDefinition>
+                    <ColumnDefinition></ColumnDefinition>
+                </Grid.ColumnDefinitions>
+                <Button x:Name="btn_Do"
+                        Click="btn_Do_Click"
+                        Grid.Column="0"
+                        Style="{StaticResource MaterialDesignOutlinedButton}">Thêm</Button>
+                <Button Grid.Column="1"
+                        Style="{StaticResource MaterialDesignOutlinedButton}"
+                        Click="ButtonBase_OnClick">Đóng</Button>
+            </Grid>
+        </StackPanel>
+    </Grid>
+</Window>
