+++ ComputerShop/AdminWindow/UserController.xaml.cs
@@ -15,6 +15,7 @@ using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
 using Dashboard;
+using Dashboard.AdminWindow.Add;
 using Dashboard.Common;
 using Dashboard.Common.ViewModel;
 using Dashboard.Data.Entities;
@@ -48,104 +49,104 @@ namespace Dashboard.AdminWindow
             dtg_User.ItemsSource = user;
         }
 
-        private Result<string> AddUser()
-        {
-            var cID = Db.Context.AppUsers.FirstOrDefault(x => x.ID == tbx_ID.Text);
-            if (cID != null)
-                return new ResultError<string>("Trùng ID");
-            var cUser = Db.Context.AppUsers.FirstOrDefault(x => x.Username == tbx_Username.Text);
-            if (cUser != null)
-                return new ResultError<string>("Tài khoản đã tồn tại");
-            if (tbx_Password.Text != tbx_ConfirmPassword.Text)
-                return new ResultError<string>("Mật khẩu xác thực không đúng");
-            var user = new AppUser
-            {
-                ID = tbx_ID.Text,
-                Email = tbx_Email.Text,
-                Dob = dp_DoB.SelectedDate ?? DateTime.Today,
-                FirstName = tbx_FirstName.Text,
-                LastName = tbx_LastName.Text,
-                Username = tbx_Username.Text,
-                PhoneNumber = tbx_PhoneNumber.Text
-                
-            };
-            if (tbx_Password.Text != tbx_ConfirmPassword.Text) return new ResultError<string>("xác thực password không chính xác");
-            user.PasswordHash = UserService.PasswordHash(tbx_Password.Text);
+        //private Result<string> AddUser()
+        //{
+        //    var cID = Db.Context.AppUsers.FirstOrDefault(x => x.ID == tbx_ID.Text);
+        //    if (cID != null)
+        //        return new ResultError<string>("Trùng ID");
+        //    var cUser = Db.Context.AppUsers.FirstOrDefault(x => x.Username == tbx_Username.Text);
+        //    if (cUser != null)
+        //        return new ResultError<string>("Tài khoản đã tồn tại");
+        //    if (tbx_Password.Text != tbx_ConfirmPassword.Text)
+        //        return new ResultError<string>("Mật khẩu xác thực không đúng");
+        //    var user = new AppUser
+        //    {
+        //        ID = tbx_ID.Text,
+        //        Email = tbx_Email.Text,
+        //        Dob = dp_DoB.SelectedDate ?? DateTime.Today,
+        //        FirstName = tbx_FirstName.Text,
+        //        LastName = tbx_LastName.Text,
+        //        Username = tbx_Username.Text,
+        //        PhoneNumber = tbx_PhoneNumber.Text
 
-            Db.Context.AppUsers.Add(user);
+        //    };
+        //    if (tbx_Password.Text != tbx_ConfirmPassword.Text) return new ResultError<string>("xác thực password không chính xác");
+        //    user.PasswordHash = UserService.PasswordHash(tbx_Password.Text);
 
-            var userRole = new AppUserRole();
-            if (tbtn_IsAdmin.IsChecked == true)
-            {
-                userRole.UserID = tbx_ID.Text;
-                userRole.RoleID = "admin";
-            }
-            else
-            {
-                userRole.UserID = tbx_ID.Text;
-                userRole.RoleID = "staff";
-            }
-            Db.Context.AppUserRoles.Add(userRole);
+        //    Db.Context.AppUsers.Add(user);
 
-            Db.Context.SaveChanges();
+        //    var userRole = new AppUserRole();
+        //    if (tbtn_IsAdmin.IsChecked == true)
+        //    {
+        //        userRole.UserID = tbx_ID.Text;
+        //        userRole.RoleID = "admin";
+        //    }
+        //    else
+        //    {
+        //        userRole.UserID = tbx_ID.Text;
+        //        userRole.RoleID = "staff";
+        //    }
+        //    Db.Context.AppUserRoles.Add(userRole);
 
-            return new ResultSuccess<string>();
-        }
+        //    Db.Context.SaveChanges();
 
-        private void SetValueUser()
-        {
-            var info = (AppUser)dtg_User.SelectedValue;
-            tbx_ID.Text = info.ID;
-            tbx_LastName.Text = info.LastName;
-            tbx_FirstName.Text = info.FirstName;
-            dp_DoB.Text = $"{info.Dob}";
-            tbx_Email.Text = info.Email;
-            tbx_PhoneNumber.Text = info.PhoneNumber;
-            tbx_Username.Text = info.Username;
-            
-            var role = Db.Context.AppUserRoles.FirstOrDefault(x => x.UserID == info.ID);
+        //    return new ResultSuccess<string>();
+        //}
 
-            tbtn_IsAdmin.IsChecked = role != null && (role.RoleID == "admin" ? true : false);
+        //private void SetValueUser()
+        //{
+        //    var info = (AppUser)dtg_User.SelectedValue;
+        //    tbx_ID.Text = info.ID;
+        //    tbx_LastName.Text = info.LastName;
+        //    tbx_FirstName.Text = info.FirstName;
+        //    dp_DoB.Text = $"{info.Dob}";
+        //    tbx_Email.Text = info.Email;
+        //    tbx_PhoneNumber.Text = info.PhoneNumber;
+        //    tbx_Username.Text = info.Username;
 
-        }
+        //    var role = Db.Context.AppUserRoles.FirstOrDefault(x => x.UserID == info.ID);
 
-        private Result<string> EditUser()
-        {
-            var user = Db.Context.AppUsers.FirstOrDefault(x => x.Username == tbx_Username.Text);
-            if (user == null)
-                return new ResultError<string>("Không có tài khoản này");
-            user.ID = tbx_ID.Text;
-            user.Email = tbx_Email.Text;
-            user.Dob = dp_DoB.SelectedDate ?? DateTime.Today;
-            user.FirstName = tbx_FirstName.Text;
-            user.LastName = tbx_LastName.Text;
-            user.Username = tbx_Username.Text;
-            user.PhoneNumber = tbx_PhoneNumber.Text;
-            if (!string.IsNullOrEmpty(tbx_Password.Text))
-            {
-                if (tbx_Password.Text != tbx_ConfirmPassword.Text) return new ResultError<string>("xác thực password không chính xác");
-                user.PasswordHash = UserService.PasswordHash(tbx_Password.Text);
+        //    tbtn_IsAdmin.IsChecked = role != null && (role.RoleID == "admin" ? true : false);
 
-            }
+        //}
 
-            var userRole = Db.Context.AppUserRoles.FirstOrDefault(x => x.UserID == user.ID);
-            if (userRole != null)
-            {
-                if (tbtn_IsAdmin.IsChecked == true)
-                {
-                    userRole.UserID = tbx_ID.Text;
-                    userRole.RoleID = "admin";
-                }
-                else
-                {
-                    userRole.UserID = tbx_ID.Text;
-                    userRole.RoleID = "staff";
-                }
-            }
+        //private Result<string> EditUser()
+        //{
+        //    var user = Db.Context.AppUsers.FirstOrDefault(x => x.Username == tbx_Username.Text);
+        //    if (user == null)
+        //        return new ResultError<string>("Không có tài khoản này");
+        //    user.ID = tbx_ID.Text;
+        //    user.Email = tbx_Email.Text;
+        //    user.Dob = dp_DoB.SelectedDate ?? DateTime.Today;
+        //    user.FirstName = tbx_FirstName.Text;
+        //    user.LastName = tbx_LastName.Text;
+        //    user.Username = tbx_Username.Text;
+        //    user.PhoneNumber = tbx_PhoneNumber.Text;
+        //    if (!string.IsNullOrEmpty(tbx_Password.Text))
+        //    {
+        //        if (tbx_Password.Text != tbx_ConfirmPassword.Text) return new ResultError<string>("xác thực password không chính xác");
+        //        user.PasswordHash = UserService.PasswordHash(tbx_Password.Text);
 
-            Db.Context.SaveChanges();
-            return new ResultSuccess<string>();
-        }
+        //    }
+
+        //    var userRole = Db.Context.AppUserRoles.FirstOrDefault(x => x.UserID == user.ID);
+        //    if (userRole != null)
+        //    {
+        //        if (tbtn_IsAdmin.IsChecked == true)
+        //        {
+        //            userRole.UserID = tbx_ID.Text;
+        //            userRole.RoleID = "admin";
+        //        }
+        //        else
+        //        {
+        //            userRole.UserID = tbx_ID.Text;
+        //            userRole.RoleID = "staff";
+        //        }
+        //    }
+
+        //    Db.Context.SaveChanges();
+        //    return new ResultSuccess<string>();
+        //}
 
         private Result<string> RemoveUser()
         {
@@ -156,7 +157,7 @@ namespace Dashboard.AdminWindow
                 return new ResultError<string>("Tài khoản không tồn tại");
             }
             Db.Context.AppUsers.Remove(user);
-            return new ResultSuccess<string>("Xóa thành công");;
+            return new ResultSuccess<string>("Xóa thành công"); ;
         }
 
         private void NumberValidationTextBox(object sender, TextCompositionEventArgs e)
@@ -167,9 +168,13 @@ namespace Dashboard.AdminWindow
 
         private void btn_AddUser_Click(object sender, RoutedEventArgs e)
         {
-            dlh_Loading.IsOpen = true;
-            _dialog = St.Add;
-            btn_Do.Content = "Thêm";
+            var au = new AddUserPage(dtg_User, AddUserPage.St.Add);
+            
+            if(au.ShowDialog()==true)
+                LoadData();
+            //dlh_Loading.IsOpen = true;
+            //_dialog = St.Add;
+            //btn_Do.Content = "Thêm";
         }
 
 
@@ -177,12 +182,14 @@ namespace Dashboard.AdminWindow
         {
             if(dtg_User.SelectedIndex < 0)
                 return;
-            _dialog = St.Edit;
-            
-            SetValueUser();
-            
-            dlh_Loading.IsOpen = true;
-            btn_Do.Content = "Sửa";
+            var au = new AddUserPage(dtg_User, AddUserPage.St.Edit){tbx_ID = {IsReadOnly = true}, btn_Do = {Content = "Edit"}};
+            au.ShowDialog();
+            //_dialog = St.Edit;
+
+            //SetValueUser();
+
+            //dlh_Loading.IsOpen = true;
+            //btn_Do.Content = "Sửa";
         }
 
         private void Btn_AddRemove_OnClick(object sender, RoutedEventArgs e)
@@ -201,53 +208,53 @@ namespace Dashboard.AdminWindow
 
         private void btn_Do_Click(object sender, RoutedEventArgs e)
         {
-            switch (_dialog)
-            {
-                case St.None:
-                    return;
-                case St.Add:
-                    var add = AddUser();
-                    if (add.IsSuccessed == false)
-                    {
-                        var mess = new MessageDialog()
-                        {
-                            tbl_Title = { Text = "LỖI" }
-                            ,
-                            tbl_Message = { Text = add.Message}
-                        };
-                        mess.ShowDialog();
-                    }
+            //switch (_dialog)
+            //{
+            //    case St.None:
+            //        return;
+            //    case St.Add:
+            //        var add = AddUser();
+            //        if (add.IsSuccessed == false)
+            //        {
+            //            var mess = new MessageDialog()
+            //            {
+            //                tbl_Title = { Text = "LỖI" }
+            //                ,
+            //                tbl_Message = { Text = add.Message}
+            //            };
+            //            mess.ShowDialog();
+            //        }
                         
-                    break;
-                case St.Edit:
-                    var edit = EditUser();
-                    if (edit.IsSuccessed == false)
-                    {
-                        if (edit.IsSuccessed == false)
-                        {
-                            var mess = new MessageDialog()
-                            {
-                                tbl_Title = { Text = "LỖI" }
-                                ,
-                                tbl_Message = { Text = edit.Message }
-                            };
-                            mess.ShowDialog();
-                        }
-                    }
-                    break;
-                default:
-                    return;
-            }
+            //        break;
+            //    case St.Edit:
+            //        var edit = EditUser();
+            //        if (edit.IsSuccessed == false)
+            //        {
+            //            if (edit.IsSuccessed == false)
+            //            {
+            //                var mess = new MessageDialog()
+            //                {
+            //                    tbl_Title = { Text = "LỖI" }
+            //                    ,
+            //                    tbl_Message = { Text = edit.Message }
+            //                };
+            //                mess.ShowDialog();
+            //            }
+            //        }
+            //        break;
+            //    default:
+            //        return;
+            //}
             _dialog = St.None;
-            btn_Do.Content = "Do";
-            dlh_Loading.IsOpen = false;
+            //btn_Do.Content = "Do";
+            //dlh_Loading.IsOpen = false;
             LoadData();
         }
 
 
         private void btn_OpenCalendar_Click(object sender, RoutedEventArgs e)
         {
-            dp_DoB.IsDropDownOpen = true;
+            //dp_DoB.IsDropDownOpen = true;
 
         }
     }
