+++ ComputerShop/Dashboard.csproj
@@ -4,7 +4,7 @@
     <OutputType>WinExe</OutputType>
     <TargetFramework>netcoreapp3.1</TargetFramework>
     <UseWPF>true</UseWPF>
-    <ApplicationIcon>icon_laptop_rec.ico</ApplicationIcon>
+    <ApplicationIcon>logo_laptop_1HI_icon.ico</ApplicationIcon>
   </PropertyGroup>
 
   <ItemGroup>
@@ -18,6 +18,8 @@
 
   <ItemGroup>
     <None Remove="0dc9da469ec12fb17c17fb856a884390.jpg" />
+    <None Remove="Login\image\WallpaperLaptopLogo.png" />
+    <None Remove="Login\WallpaperLaptopLogo.png" />
     <None Remove="Staff\AORUS-Z390-Xtreme-Waterforce-5G-Bundle-Intel-Core-i9-9900K.jpg" />
     <None Remove="Staff\aorus_rgb_memory16gb.jpg" />
     <None Remove="Staff\AORUS_Z390_Xtreme_Waterforce_5G_Bundle_Intel_Core_i9_9900K.jpg" />
@@ -61,6 +63,7 @@
     <Resource Include="Login\image\Loading.gif" />
     <Resource Include="Login\image\loginwpf.png" />
     <Resource Include="Login\image\wallpaper2.png" />
+    <Resource Include="Login\image\WallpaperLaptopLogo.png" />
     <Resource Include="Login\image\WallpaperVTT.png" />
     <Resource Include="Staff\aorus_rgb_memory16gb.jpg" />
     <Resource Include="Staff\AORUS_Z390_Xtreme_Waterforce_5G_Bundle_Intel_Core_i9_9900K.jpg" />
